fastapi/requirements.txt
anyio==3.6.1
autopep8==1.7.0
bcrypt==4.0.0
beanie==1.11.9
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.83.0
fastapi-users==10.1.5
fastapi-users-db-beanie==1.1.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
httpx-oauth==0.10.1
idna==3.3
makefun==1.15.0
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
yarl==1.8.1

fastapi-users[beanie]
tortoise-orm
databases[postgresql]

python-dateutil
Pillow
numpy
fastapi-mail